+++ widgets/buttonSelector.js
@@ -135,9 +135,16 @@ module.exports = {
                     element.cssEditor.destroy();
                     delete element.cssEditor;
                 `,
-            //When the user clicks the "reset configuration" set the options to their defaults
-            reset: `
-                    element.cssEditor.setValue(defaultConfig.CSS.value);
+            //When the user clicks the "copy configuration" button update the values shown
+            update: `
+                    var optionsList = $("#options");
+                    optionsList.editableList("empty");
+                    for(var i in settings.options.value) {
+                        var option = settings.options.value[i];
+                        optionsList.editableList('addItem', { label: option.label, value: option.value, onColor: option.onColor, offColor: option.offColor });
+                    }
+                
+                    element.cssEditor.setValue(settings.CSS.value);
                     element.cssEditor.clearSelection();
                 `
         }