+++ CHANGELOG.md
@@ -12,3 +12,6 @@
 ## - 0.2.2
 * Fixed a few issues with the volume widget
 * Fixed a bug where the width and height multipliers were not being set
+## - 0.2.4
+* More changes to the volume widget
+* Added the ability to copy settings from another widget